重構工具架構
