+++ src/tools/thought/index.ts
@@ -1,0 +1,4 @@
+// 導出所有思維鏈工具
+
+// processThought
+export { processThought, processThoughtSchema } from "./processThought.js";
